Refactor update to v2
